comment 4

+++ amqp_lib/decode.c
@@ -3,8 +3,8 @@
 #include <types.h>
 
 
-int decodeDescribedFormatCode(char *source_buffer, PDescribedFormatCode out_described_format_code) {
-    int offset = 0;
+size_t decodeDescribedFormatCode(char *source_buffer, PDescribedFormatCode out_described_format_code) {
+    size_t offset = 0;
     offset += read_char(source_buffer, &out_described_format_code->v1);
     offset += read_char(source_buffer + offset, &out_described_format_code->v2);
     offset += read_char(source_buffer + offset, &out_described_format_code->formatCode);
@@ -12,8 +12,8 @@ int decodeDescribedFormatCode(char *source_buffer, PDescribedFormatCode out_desc
     return 0;
 }
 
-int parseAmqp10MessageBuffer(char *source_buffer, long len, PMessage_t out_message) {
-    int offset = 0;
+size_t parseAmqp10MessageBuffer(char *source_buffer, long len, PMessage_t out_message) {
+    size_t offset = 0;
     while (offset != len) {
         DescribedFormatCode formatCode;
         decodeDescribedFormatCode(source_buffer, &formatCode);
